* emacstv.el: Adding Emacs Compatibilities | Doc Refinement
Adding important Documentation for functions in Emacs and fixing some
miss-declarations/requirements to avoid unpredictable behaviour and
conflicts.

Improving the regexp search URL for ~YouTube~ Videos to avoid multiple
conditions for this.

TODO: I'll open an Issue to propose the creation of a document to have
Code Guidelines for identation and code documentation if you are agree with
that. This because that will help us understand better the code.

Also I'll open another Issue to handled the regexp searchs in the
videos.org, with mission to improve performance and avoid unpredictable behaviours.

Note: I tried to avoid format all the Document to not mess up with all
the other things, however there are some lines that I couldn't format to
just be capable of understand the code.

Modified-by: Cristian D. Moreno - Kyonax <iam@kyo.wtf>